+++ samples/todomvc/java/fletch/BoxedString.java
@@ -31,9 +31,9 @@ public class BoxedString extends Reader {
   }
 
   public String getStr() { return readString(0); }
-  public Uint8List getStrData() {
+  public Uint16List getStrData() {
     ListReader reader = new ListReader();
     readList(reader, 0);
-    return new Uint8List(reader);
+    return new Uint16List(reader);
   }
 }
